Add unhandled exception logging & tracking for Scheduled Tasks
I noticed in the logs that the Stale letters task was failing in our
E2E tests and encountering an exception, but this did not fail the test
because there is no tracking of exceptions that occur in tasks.

The new ThreadPoolConfig class adds a custom error handler that
will log any unhandled exception thrown from a task and increment a
counter.

Our integration tests can then make assertions against this error count
to verify no unexpected errors occured.

This change also fixes the Stale letters task, which was failing due to
missing a @Transactional attribute - hibernate queries that return
streams must run within a transaction.