Fix race condition in end to end tests
Adds a final check to end to end tests to wait for the Xerox reports
to be deleted by the MarkLettersPosted task.

This ensures that we do not shut down the SFTP server whilst the job
is still running, which would cause an exception when the job tried
to delete the xerox report but the sftp server had shut down.