Use pagination instead of streams for upload letters
Hibernate's streams must be called from within a transaction.

This is undesirable for batch processing tasks that process large
numbers of entities, since the transaction will be rolled back if there
is a problem anywhere in the job, even for entities that were processed
successfully.

This change uses spring's pagination to process letters in batches.