Upgrade SSHD test SFTP server to 1.7.0
This fixes the intermittent test failure 'KeyExchange signature
verification failed', which I believe is due to a bug in the really
old version of sshd-sftp we were using, instead of the up to date sftp
server in sshd-core.

To investigate this I set up a test loop where our FtpClient continually
 connected and disconnected to a local sshd server.

With the old version of sshd-sftp I would get a KeyExchange
exception after a few seconds to a couple of minutes. After upgrading to
sshd-core 1.7.0 I was able to run the loop for an hour without any
 problems.